requirements.txt
cPython
lightgbm
pexpect
matplotlib
numpy
scipy
scikit-learn
pandas
termcolor
contextlib2
prettytable
graphviz
configparser
pillow
pytest
pyyaml
requests
setuptools
pycocotools
opencv-python==4.3.0.36
imgviz
scikit-image
pydicom
imageio
labelme
thop
flask
pyradiomics
xgboost
seaborn
nibabel
pynrrd
moviepy
lxml
SimpleITK
statsmodels
imblearn
joblib